homework8

+++ src/news/comments/comments.service.ts
@@ -1,27 +1,27 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
-//import { News } from '../news.interface';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CommentsEntity } from './comments.entity';
 import { Repository } from 'typeorm';
 import { NewsService } from '../news.service';
 import { UsersService } from '../../users/users.service';
+import { EventsComment } from './EventsComment.enum';
+import {
+  checkPermission,
+  Modules,
+} from '../../auth/role/utils/check-permission';
 
 export type Comment = {
   id?: number;
   message: string;
   author: string;
-  createdAt?: Date;
-  idNews: number;
 };
 
 export type CommentEdit = {
   id?: number;
   message?: string;
   author?: string;
-  createdAt?: Date;
-}
-
+};
 @Injectable()
 export class CommentsService {
   constructor(
@@ -31,8 +31,6 @@ export class CommentsService {
     private readonly userService: UsersService,
     private readonly eventEmitter: EventEmitter2,
   ) {}
-  private readonly comments = [];
-
   async create(
     idNews: number,
     message: string,
@@ -57,33 +55,37 @@ export class CommentsService {
         },
         HttpStatus.NOT_FOUND,
       );
-  }
-
-  const commentEntity = new CommentsEntity();
-  commentEntity.news = _news;
-  commentEntity.user = _user;
-  commentEntity.message = message;
-    
-  return this.commentsRepository.save(commentEntity);
     }
-
-    async edit(idComment: number, comment: CommentEdit): Promise<CommentsEntity> {
-      const _comment = await this.commentsRepository.findOne(idComment);
-      _comment.message = comment.message;
-      return this.commentsRepository.save(_comment);
+    const commentEntity = new CommentsEntity();
+    commentEntity.news = _news;
+    commentEntity.user = _user;
+    commentEntity.message = message;
+    return this.commentsRepository.save(commentEntity);
   }
-
+  async edit(idComment: number, comment: CommentEdit): Promise<CommentsEntity> {
+    const _comment = await this.commentsRepository.findOne({
+      where: { id: idComment },
+      relations: ['news', 'user'],
+    });
+    _comment.message = comment.message;
+    const _updatedComment = await this.commentsRepository.save(_comment);
+    this.eventEmitter.emit(EventsComment.edit, {
+      commentId: idComment,
+      newsId: _comment.news.id,
+      comment: _updatedComment,
+    });
+    return _updatedComment;
+  }
   async findAll(idNews: number): Promise<CommentsEntity[]> {
     return this.commentsRepository.find({
       where: { news: idNews },
       relations: ['user'],
     });
   }
-
-  async remove(idComment: number): Promise<CommentsEntity> {
+  async remove(idComment: number, userId: number): Promise<CommentsEntity> {
     const _comment = await this.commentsRepository.findOne({
       where: { id: idComment },
-      relations: ['news'],
+      relations: ['news', 'user'],
     });
     if (!_comment) {
       throw new HttpException(
@@ -95,18 +97,28 @@ export class CommentsService {
       );
     }
 
+    const _user = await this.userService.findById(userId);
+    if (
+      _user.id !== _comment.user.id &&
+      !checkPermission(Modules.editComment, _user.roles)
+    ) {
+      throw new HttpException(
+        {
+          status: HttpStatus.FORBIDDEN,
+          error: 'Недостаточно прав для удаления',
+        },
+        HttpStatus.FORBIDDEN,
+      );
+    }
     const comment = await this.commentsRepository.remove(_comment);
-    this.eventEmitter.emit('comment.remove', {
+    this.eventEmitter.emit(EventsComment.remove, {
       commentId: idComment,
       newsId: _comment.news.id,
     });
-
     return comment;
   }
-
   async removeAll(idNews) {
     const _comments = await this.findAll(idNews);
     return await this.commentsRepository.remove(_comments);
-  }  
-
+  }
 }
